requirements.txt
openai 
fastapi 
uvicorn 
python-dotenv 
websockets